+++ feature/src/main/java/com/gdg/feature/map/MapSideEffect.kt
@@ -1,5 +1,5 @@
 package com.gdg.feature.map
 
 sealed class MapSideEffect {
-    data class NavigateToDetail(val id: Int) : MapSideEffect()
+    data class NavigateToDetail(val id: Long) : MapSideEffect()
 }
